requirements.txt
pyside6
cryptography
argon2-cffi
pynput
